refectoring, and warning cleanup
