--- src/Carousel.js
@@ -1,10 +1,0 @@
-
-function Carousel() {
-  return (
-    <div className="carousel-container">
-      
-    </div>
-  );
-}
-
-export default Carousel;
