--- src/components/ArticleCarousel/carouselContent.js
@@ -1,20 +1,0 @@
-export const articles = [
-    {
-    title: "What it's like when racism comes for you",
-    url: "https://www.theatlantic.com/politics/archive/2021/03/anti-asian-harassments-long-history-america/618211/",
-    imageUrl: "./assets/article-slide-01.png",
-    imageAlt: "Hate is a Virus"
-    },
-    {
-    title: "How to help stop hate crimes against Asian Americans",
-    url: "https://www.papermag.com/help-stop-hate-crimes-asian-americans-2651116989.html?rebelltitem=24#rebelltitem24",
-    imageUrl: "./assets/article-slide-02.png",
-    imageAlt: "Stop Asian Hate"
-    },
-    {
-    title: "The History of Anti-Asian Sentiment in the U.S.",
-    url: "https://www.npr.org/transcripts/978832077",
-    imageUrl: "./assets/article-slide-03.png",
-    imageAlt: "Anti-Chinese labor meeting on March 20, 1880 outside of San Francisco City Hall"
-    }
-]
