+++ src/components/Successes/Successes.js
@@ -1,34 +1,27 @@
 import './Successes.css'
 import Carousel from 'react-bootstrap/Carousel'
+import {successes} from "../../carouselContent"
 
 
 const Successes = () => {
 
+  const successContent = successes.map((success, index) => 
+    <Carousel.Item key={index}>
+        <img
+        className="d-block w-100"
+        src={success.imageUrl}
+        alt={success.imageAlt}
+        />
+        <Carousel.Caption key={index}>
+            <a href={success.url} target="_blank" rel="noreferrer"><h3>{success.title}</h3></a>
+        </Carousel.Caption>
+    </Carousel.Item>
+    )
+
   return (
     <section className="successes-container row">
       <Carousel>
-        <Carousel.Item>
-          <img
-            className="d-block w-100"
-            src="holder.js/800x400?text=First slide&bg=373940"
-            alt="First slide"
-          />
-          <Carousel.Caption>
-            <h3>First slide label</h3>
-            <p>Nulla vitae elit libero, a pharetra augue mollis interdum.</p>
-          </Carousel.Caption>
-        </Carousel.Item>
-        <Carousel.Item>
-          <img
-            className="d-block w-100"
-            src="holder.js/800x400?text=First slide&bg=373940"
-            alt="Secondslide"
-          />
-          <Carousel.Caption>
-            <h3>Second slide label</h3>
-            <p>Nulla vitae elit libero, a pharetra augue mollis interdum.</p>
-          </Carousel.Caption>
-        </Carousel.Item>
+          {successContent}
       </Carousel>
     </section>
   )
